refactor: rename authorization module to auth
